requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+git@github.com:UiL-OTS-labs/django-shared-core.git@master#egg=uil-django-core
asn1crypto==0.24.0        # via cryptography
blessings==1.7            # via curtsies
bpython==0.17.1
certifi==2018.11.29       # via requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
cryptography==2.4.2
curtsies==0.3.0           # via bpython
django-braces==1.13.0
django-csp==3.5
django-debug-toolbar==1.11
django-simple-menu==1.2.1
django==2.0.10
greenlet==0.4.15          # via bpython
idna==2.8                 # via cryptography, requests
pillow==5.4.1
pip-tools==3.2.0
pycparser==2.19           # via cffi
pygments==2.3.1           # via bpython
pyjwt==1.7.1
pytz==2018.9              # via django
requests==2.21.0
six==1.12.0               # via blessings, bpython, cryptography, pip-tools
sqlparse==0.2.4           # via django-debug-toolbar
typing==3.6.6             # via curtsies
urllib3==1.24.1           # via requests
wcwidth==0.1.7            # via curtsies
